text field: extract layout to own composable helper
